requirements.txt
openai-whisper==20231117
python-docx==0.8.11
numpy==1.24.3
Pillow>=10.0.0
tkinterdnd2>=0.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
openai>=1.0.0
